+++ Holovibes/sources/compute/preprocessing.cc
@@ -63,7 +63,7 @@ namespace holovibes
 
 		float Preprocessing::compute_current_intensity(cufftComplex* buffer_ptr, size_t nb_pixels)
 		{
-			float res = average_operator_from_complex(buffer_ptr, nb_pixels);
+			float res = average_operator_from_complex(buffer_ptr, static_cast<uint>(nb_pixels));
 			cudaStreamSynchronize(0);
 			return res;
 		}
